+++ Servicios/GeneradorBase.cs
@@ -5,10 +5,10 @@ namespace SimulacionTP1.Servicios
     public abstract class GeneradorBase
     {
 
-        public double[] Generar(ulong semilla, ulong modulo, ulong a, ulong c)
+        public double[] Generar(int semilla, int modulo, int a, int c)
         {
             double[] nums = new double[20];
-            ulong xi = semilla;
+            int xi = semilla;
 
             for (int i = 0; i < nums.Length; i++)
             {
@@ -19,12 +19,12 @@ namespace SimulacionTP1.Servicios
             return nums;
         }
 
-        public double GenerarSiguiente(ulong x, ulong modulo, ulong a, ulong c)
+        public double GenerarSiguiente(int x, int modulo, int a, int c)
         {
             return (double) (Calcular(x, a, c) % modulo) / modulo;
         }
 
-        public void Validar(ulong x0, ulong m, ulong a, ulong c)
+        public void Validar(int x0, int m, int a, int c)
         {
             if (x0 <= 0) throw new ApplicationException("Valor de semilla inválido");
             if (m <= 0) throw new ApplicationException("Valor de módulo inválido");
@@ -32,7 +32,7 @@ namespace SimulacionTP1.Servicios
             ValidarC(c);
         }
 
-        public string CalcularM(ulong g)
+        public string CalcularM(int g)
         {
             return Math.Pow(2, g).ToString();
         }
@@ -59,8 +59,8 @@ namespace SimulacionTP1.Servicios
         //    }
         //}
 
-        protected abstract void ValidarC(ulong c);
-        protected abstract ulong Calcular(ulong x, ulong a, ulong c);
-        public abstract string CalcularA(ulong k);
+        protected abstract void ValidarC(int c);
+        protected abstract int Calcular(int x, int a, int c);
+        public abstract string CalcularA(int k);
     }
 }
