+++ Servicios/GeneradorMultiplicativo.cs
@@ -2,17 +2,17 @@
 {
     public class GeneradorMultiplicativo : GeneradorBase
     {
-        protected override ulong Calcular(ulong x, ulong a, ulong c)
+        protected override int Calcular(int x, int a, int c)
         {
             return a * x;
         }
 
-        public override string CalcularA(ulong k)
+        public override string CalcularA(int k)
         {
             return (3 + 8 * k).ToString();
         }
 
-        protected override void ValidarC(ulong c)
+        protected override void ValidarC(int c)
         {
             
         }
